style: apply safe clippy idioms across non-rpa crates
Behavior-preserving cleanups, each verified by the crate's tests:

- assign-ops (x = x op y → x op= y): uhf/rohf Fock builds (f_a -= k…,
  f += shift…), CCD/CCSD amplitude updates (t1/t2 += delta)
- redundant same-type casts dropped: ao_grid radial() `shell.l as i32`
  (already i32), u_oo_rimp2 `mol.nelec() as i32` (already i32)
- needless borrows removed: &rhf.mos_r()/&rhf.eps_r()/&result.density_r()
  (&&T → &T) in mp2/scf gradients and rimp2; &self.prep (already &T) in
  link_k, matching the existing self.prep call site
- becke a_corr size-correction: if>/if< pair → .clamp(-0.5, 0.5)
  (identical for finite/inf/NaN inputs)
- qqr decay: single-arm match + `_ => {}` → `if ==` (OperatorKind: PartialEq)
- link_k bitvec sizing: (n+63)/64 → n.div_ceil(64)
- RiMp2Config: hand-written Default → #[derive(Default)]
- fxc: useless format!("literal") → "literal".to_string()
- minimax table lookup: `&(.., ref t, ref w)` → match-ergonomics `(.., t, w)`

Verified: warning-free `cargo build --workspace`; ferric-cc 3/3,
ferric-quadrature 2/2, ferric-scf lib 37/37, ferric-dft becke 7/7,
ferric-mp2 19/19 (incl. oo/u-oo/u orbital-gradient-vs-FD).

+++ crates/ferric-dft/src/ao_grid.rs
@@ -70,7 +70,7 @@ pub fn nbasis(mol: &Molecule, bs: &BasisSet) -> Result<usize, GtoEvalError> {
 /// even with the wrong absolute scale).
 #[inline]
 fn radial(shell: &LocatedShell, r2: f64) -> f64 {
-    let l = shell.l as i32;
+    let l = shell.l;
     let pi = std::f64::consts::PI;
     // (2l-1)!! for l = 0..3
     let dbl_fact: f64 = match l {
@@ -220,7 +220,7 @@ pub fn eval_basis_on_points(
 /// Then ∂(rad)/∂x = 2x · d_rad/d(r²).
 #[inline]
 fn radial_and_d(shell: &LocatedShell, r2: f64) -> (f64, f64) {
-    let l = shell.l as i32;
+    let l = shell.l;
     let pi = std::f64::consts::PI;
     // (2l-1)!! for l = 0..3
     let dbl_fact: f64 = match l {
@@ -243,7 +243,7 @@ fn radial_and_d(shell: &LocatedShell, r2: f64) -> (f64, f64) {
 /// Like `radial_and_d` but also returns d²rad/d(r²)² = Σ c·N·α²·exp(-α r²).
 #[inline]
 fn radial_and_d_d2(shell: &LocatedShell, r2: f64) -> (f64, f64, f64) {
-    let l = shell.l as i32;
+    let l = shell.l;
     let pi = std::f64::consts::PI;
     let dbl_fact: f64 = match l {
         0 | 1 => 1.0,
